requirements.txt
alembic==1.17.2
fastapi==0.128.0
passlib==1.7.4
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
uvicorn==0.40.0
openai>=1.0.0
argon2-cffi==23.1.0
asyncpg==0.29.0
streamlit==1.52.2
requests==2.31.0
pandas==2.3.3